backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
attrs==24.2.0
azure-cognitiveservices-speech==1.40.0
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
deprecation==2.1.0
dnspython==2.6.1
edge-tts==6.1.13
email_validator==2.2.0
frozenlist==1.4.1
gotrue==2.8.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5tagger==1.3.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
postgrest==0.16.11
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.0.2
requests==2.32.3
sanic==24.6.0
Sanic-Cors==2.2.0
sanic-ext==23.12.0
sanic-routing==23.12.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
storage3==0.7.7
StrEnum==0.4.15
supafunc==0.5.1
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvloop==0.20.0
websockets==12.0
yarl==1.11.1
